% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{check_raw_counts}
\alias{check_raw_counts}
\title{Validate raw count table}
\usage{
check_raw_counts(
  counts,
  gene_col = c("gene_id", "GeneID", "gene", "Gene", "ENSEMBL", "ensembl"),
  feature_col = c("gene_name", "SYMBOL", "symbol"),
  min_library_size = 1e+06,
  min_detected_genes = 5000
)
}
\arguments{
\item{counts}{Raw count table.}

\item{gene_col}{Possible gene ID columns.}

\item{feature_col}{Possible gene symbol columns.}

\item{min_library_size}{Minimum acceptable library size.}

\item{min_detected_genes}{Minimum acceptable detected genes.}
}
\value{
List of QC results.
}
\description{
Performs sanity checks on bulk RNA-seq count tables.
}
