% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{deggo_extract_deg_genes}
\alias{deggo_extract_deg_genes}
\title{Extract selected DEG genes from DEGgo results}
\usage{
deggo_extract_deg_genes(results, genes)
}
\arguments{
\item{results}{DEGgo result object.}

\item{genes}{Character vector of gene symbols.}
}
\value{
A data.frame.
}
\description{
Extract selected DEG genes from DEGgo results
}
