% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{deggo_extract_go_genes}
\alias{deggo_extract_go_genes}
\title{Extract genes found in GO terms}
\usage{
deggo_extract_go_genes(go_results, genes, gene_sep = "/")
}
\arguments{
\item{go_results}{GO enrichment result data.frame.}

\item{genes}{Character vector of genes to search.}

\item{gene_sep}{Separator used in the geneID column.}
}
\value{
A data.frame with GO terms containing selected genes.
}
\description{
Extract genes found in GO terms
}
