% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{deggo_extract_go_genes_pairwise}
\alias{deggo_extract_go_genes_pairwise}
\title{Extract selected genes from pairwise GO results}
\usage{
deggo_extract_go_genes_pairwise(results, genes)
}
\arguments{
\item{results}{DEGgo result object.}

\item{genes}{Character vector of genes.}
}
\value{
A data.frame.
}
\description{
Extract selected genes from pairwise GO results
}
