% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{deggo_extract_go_keywords}
\alias{deggo_extract_go_keywords}
\title{Extract GO terms matching keywords}
\usage{
deggo_extract_go_keywords(results, keywords)
}
\arguments{
\item{results}{DEGgo result object.}

\item{keywords}{Character vector of keywords.}
}
\value{
A data.frame.
}
\description{
Extract GO terms matching keywords
}
