% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{deggo_extract_rhythmic_genes}
\alias{deggo_extract_rhythmic_genes}
\title{Extract selected genes from DEGgo rhythmicity results}
\usage{
deggo_extract_rhythmic_genes(results, genes)
}
\arguments{
\item{results}{A \code{deggo_rhythm_results} object returned by
\code{\link[=run_deggo_rhythmicity]{run_deggo_rhythmicity()}}.}

\item{genes}{Character vector of gene identifiers.}
}
\value{
A data.frame.
}
\description{
Extract selected genes from DEGgo rhythmicity results
}
