% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_expression}
\alias{extract_expression}
\title{Extract normalized gene expression}
\usage{
extract_expression(
  dds,
  metadata,
  genes,
  assay = c("vst", "normalized", "log2_normalized", "raw"),
  gene_col = "SYMBOL",
  collapse_fun = sum
)
}
\arguments{
\item{dds}{DESeqDataSet object.}

\item{metadata}{Sample metadata.}

\item{genes}{Character vector of gene symbols or identifiers.}

\item{assay}{One of \code{"vst"}, \code{"normalized"}, \code{"log2_normalized"} or \code{"raw"}.}

\item{gene_col}{Character. Column containing gene names or identifiers in rowData(dds).}

\item{collapse_fun}{Function used to collapse duplicated gene rows.}
}
\value{
Long-format expression table.
}
\description{
Extract normalized or variance-stabilized expression values from a
DESeq2 object and return a long-format table.
}
