% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_qc.R
\name{plot.deggo_design_qc}
\alias{plot.deggo_design_qc}
\title{Plot DesignQC PCA}
\usage{
\method{plot}{deggo_design_qc}(x, color_by = NULL, shape_by = NULL, label = FALSE, ...)
}
\arguments{
\item{x}{A deggo_design_qc object.}

\item{color_by}{Metadata variable used for colour.}

\item{shape_by}{Optional metadata variable used for shape.}

\item{label}{Show sample labels.}

\item{...}{Unused.}
}
\description{
Plot DesignQC PCA
}
